Move InvalidIdentifierAttributeTypeException from data to data-validation
