+++ molgenis-data-import/src/main/java/org/molgenis/data/importer/emx/EmxMetaDataParser.java
@@ -24,6 +24,7 @@ import org.molgenis.data.meta.model.*;
 import org.molgenis.data.meta.model.Package;
 import org.molgenis.data.support.EntityTypeUtils;
 import org.molgenis.data.validation.CompositeValidationResult;
+import org.molgenis.data.validation.InvalidIdentifierAttributeTypeException;
 import org.molgenis.data.validation.ValidationException;
 import org.molgenis.data.validation.meta.*;
 import org.molgenis.util.EntityUtils;
