+++ molgenis-data-validation/src/main/resources/l10n/data_validation_en.properties
@@ -15,6 +15,7 @@ V20=Unknown relation IRI ''{1}'' for tag ''{0}''
 # Various validation messages
 V21=Duplicate entity type ''{0}''
 V22=No {0} capabilities for entity {1}
+V23=Identifier is of type ''{0}''. Id attributes can only be of type 'STRING', 'INT' or 'LONG'
 # Package validation messages
 V30=Modifying system package ''{0}'' is not allowed
 V31=Package ''{0}'': identifier ''{1}'' contains illegal characters.
