+++ molgenis-data/src/main/resources/l10n/data_en.properties
@@ -5,7 +5,6 @@ D05=Unknown repository ''{0}''
 D06=Unknown repository collection ''{0}''
 D07=Unknown package ''{0}''
 D08=Unknown parent entity type''{0}''
-D09=Identifier is of type ''{0}''. Id attributes can only be of type 'STRING', 'INT' or 'LONG'
 D10=Entity type ''{0,label}'' attribute ''{1,label}'' value ''{2}'' can''t be converted to type ''{3}''
 D11a=Failed to parse attribute ''{0}'' value ''{1}'' as date. Valid date format is [YYYY-MM-DD].
 D11b=Failed to parse attribute ''{0}'' value ''{1}'' as datetime. Valid datetime format is [YYYY-MM-DDThh:mm:ssZ].
